README.md
It can be argued that there is a crisis in education due to the fact that the world has been changing at a very fast pace in the past few decades and schools have not been able to adjust and adapt to those changes fast enough.  Schools are providing a 20th century education to 21st century students in a 21st century world. We carried out a literature review concerning how K-12 education in and for the 21st century should look like —what major social changes there are and what the educational implications of these changes might be—.